Move actor stepping tests to separate file
